pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
# Check https://flit.readthedocs.io/en/latest/pyproject_toml.html for all available sections
name = "ansys-aedt-antenna-toolkits"
dynamic = ["version"]
description = "Toolkit to design antennas with Ansys AEDT."
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.support@ansys.com"},
]
maintainers = [
    {name = "PyAnsys developers", email = "pyansys.maintainers@ansys.com"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "build==0.10.0",
    "twine==4.0.2",
    "pyaedt",
    "importlib-metadata >=4.0",
    "PySide6-Essentials==6.5.1.1",
    "pyqtgraph",
    "qdarkstyle",
    "flask",
]

[project.optional-dependencies]
tests = [
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
]
doc = [
    "ansys-sphinx-theme==0.9.9",
    "numpydoc==1.5.0",
    "recommonmark==0.7.1",
    "sphinx==7.0.1",
    "sphinx-copybutton==0.5.2",
    "enum-tools[sphinx]==0.10.0",
    "sphinx-autoapi==2.1.1",
    "sphinx-autobuild==2021.3.14",
    "sphinx-autodoc-typehints==1.23.3",
    "sphinx-gallery==0.13.0",
    "sphinx-notfound-page==0.8.3",
    "nbsphinx==0.9.2",
]

[tool.flit.module]
name = "ansys.aedt.toolkits.antennas"

[project.urls]
Source = "https://github.com/pyansys/pyaedt-antenna-toolkit.git"
Tracker = "https://github.com/pyansys/pyaedt-antenna-toolkit.git/issues"
Homepage = "https://github.com/pyansys/pyaedt-antenna-toolkit.git"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 120
default_section = "THIRDPARTY"
src_paths = ["doc", "src", "tests"]

[tool.coverage.run]
source = ["ansys.pyaedt_toolkits"]

[tool.coverage.report]
show_missing = true
